add uart context
