+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 idf_component_register(
-  SRCS "src/esp_dmx.c" "src/dmx/agent.c" "src/dmx/sniffer.c" 
+  SRCS "src/esp_dmx.c" "src/dmx/sniffer.c" 
        "src/dmx/hal/uart.c" "src/rdm/controller.c" "src/rdm/responder.c"
        "src/rdm/parameters.c" "src/rdm/utils.c"
   INCLUDE_DIRS "src"
