+++ src/dmx/hal/uart.h
@@ -16,6 +16,7 @@
 
 #include "uart.h"
 #include "hal/uart_hal.h"
+#include "dmx/types.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -28,7 +29,10 @@ typedef struct dmx_uart_t dmx_uart_t;
  *
  * @param uart A pointer to a UART port.
  */
-void dmx_uart_init(uart_dev_t *uart);
+void dmx_uart_init(dmx_port_t dmx_num, int isr_flags);
+
+// TODO: docs
+void dmx_uart_deinit(dmx_port_t dmx_num);
 
 /**
  * @brief Gets the UART baud rate of the selected UART hardware.
@@ -36,7 +40,7 @@ void dmx_uart_init(uart_dev_t *uart);
  * @param uart A pointer to a UART port.
  * @return The baud rate of the UART hardware.
  */
-uint32_t dmx_uart_get_baud_rate(uart_dev_t *uart);
+uint32_t dmx_uart_get_baud_rate(dmx_port_t dmx_num);
 
 /**
  * @brief Sets the baud rate for the UART.
@@ -44,7 +48,7 @@ uint32_t dmx_uart_get_baud_rate(uart_dev_t *uart);
  * @param uart A pointer to a UART port.
  * @param baud_rate The baud rate to use.
  */
-void dmx_uart_set_baud_rate(uart_dev_t *uart, uint32_t baud_rate);
+void dmx_uart_set_baud_rate(dmx_port_t dmx_num, uint32_t baud_rate);
 
 /**
  * @brief Sets the number of bytes that the UART must receive to trigger a RX
@@ -54,7 +58,7 @@ void dmx_uart_set_baud_rate(uart_dev_t *uart, uint32_t baud_rate);
  * @param threshold The number of bytes needed to trigger an RX FIFO full
  * interrupt.
  */
-void dmx_uart_set_rxfifo_full(uart_dev_t *uart, uint8_t threshold);
+void dmx_uart_set_rxfifo_full(dmx_port_t dmx_num, uint8_t threshold);
 
 /**
  * @brief Sets the number of bytes that the UART TX FIFO must have remaining in
@@ -64,7 +68,7 @@ void dmx_uart_set_rxfifo_full(uart_dev_t *uart, uint8_t threshold);
  * @param threshold The number of bytes remaining to trigger a TX FIFO empty
  * interrupt.
  */
-void dmx_uart_set_txfifo_empty(uart_dev_t *uart, uint8_t threshold);
+void dmx_uart_set_txfifo_empty(dmx_port_t dmx_num, uint8_t threshold);
 
 /**
  * @brief Inverts or un-inverts the TX line on the UART.
@@ -72,7 +76,7 @@ void dmx_uart_set_txfifo_empty(uart_dev_t *uart, uint8_t threshold);
  * @param uart A pointer to a UART port.
  * @param invert_mask 1 to invert, 0 to un-invert.
  */
-void dmx_uart_invert_tx(uart_dev_t *uart, uint32_t invert);
+void dmx_uart_invert_tx(dmx_port_t dmx_num, uint32_t invert);
 
 /**
  * @brief Gets the level of the UART RTS line.
@@ -81,7 +85,7 @@ void dmx_uart_invert_tx(uart_dev_t *uart, uint32_t invert);
  * @return 1 if the UART RTS line is enabled (set low; read), 0 if the UART RTS
  * line is disable (set high; write).
  */
-int dmx_uart_get_rts(uart_dev_t *uart);
+int dmx_uart_get_rts(dmx_port_t dmx_num);
 
 /**
  * @brief Gets the interrupt status mask from the UART.
@@ -89,7 +93,7 @@ int dmx_uart_get_rts(uart_dev_t *uart);
  * @param uart A pointer to a UART port.
  * @return The interrupt status mask.
  */
-int dmx_uart_get_interrupt_status(uart_dev_t *uart);
+int dmx_uart_get_interrupt_status(dmx_port_t dmx_num);
 
 /**
  * @brief Enables UART interrupts using an interrupt mask.
@@ -97,7 +101,7 @@ int dmx_uart_get_interrupt_status(uart_dev_t *uart);
  * @param uart A pointer to a UART port.
  * @param mask The UART mask that is enabled.
  */
-void dmx_uart_enable_interrupt(uart_dev_t *uart, int mask);
+void dmx_uart_enable_interrupt(dmx_port_t dmx_num, int mask);
 
 /**
  * @brief Disables UART interrupts using an interrupt mask.
@@ -105,7 +109,7 @@ void dmx_uart_enable_interrupt(uart_dev_t *uart, int mask);
  * @param uart A pointer to a UART port.
  * @param mask The UART mask that is disabled.
  */
-void dmx_uart_disable_interrupt(uart_dev_t *uart, int mask);
+void dmx_uart_disable_interrupt(dmx_port_t dmx_num, int mask);
 
 /**
  * @brief Clears UART interrupts using a mask.
@@ -113,7 +117,7 @@ void dmx_uart_disable_interrupt(uart_dev_t *uart, int mask);
  * @param uart A pointer to a UART port.
  * @param mask The UART mask that is cleared.
  */
-void dmx_uart_clear_interrupt(uart_dev_t *uart, int mask);
+void dmx_uart_clear_interrupt(dmx_port_t dmx_num, int mask);
 
 /**
  * @brief Gets the current length of the bytes in the UART RX FIFO.
@@ -121,7 +125,7 @@ void dmx_uart_clear_interrupt(uart_dev_t *uart, int mask);
  * @param uart A pointer to a UART port.
  * @return The number of bytes in the UART RX FIFO.
  */
-uint32_t dmx_uart_get_rxfifo_len(uart_dev_t *uart);
+uint32_t dmx_uart_get_rxfifo_len(dmx_port_t dmx_num);
 
 /**
  * @brief Gets the level of the UART RX line.
@@ -130,7 +134,7 @@ uint32_t dmx_uart_get_rxfifo_len(uart_dev_t *uart);
  * @return The UART RX line level.
  */
 
-uint32_t dmx_uart_get_rx_level(uart_dev_t *uart);
+uint32_t dmx_uart_get_rx_level(dmx_port_t dmx_num);
 
 /**
  * @brief Reads from the UART RX FIFO.
@@ -141,7 +145,7 @@ uint32_t dmx_uart_get_rx_level(uart_dev_t *uart);
  * data.
  * @return The number of characters read.
  */
-void dmx_uart_read_rxfifo(uart_dev_t *uart, uint8_t *buf, int *size);
+void dmx_uart_read_rxfifo(dmx_port_t dmx_num, uint8_t *buf, int *size);
 
 /**
  * @brief Enables or disables the UART RTS line.
@@ -150,13 +154,13 @@ void dmx_uart_read_rxfifo(uart_dev_t *uart, uint8_t *buf, int *size);
  * @param set 1 to enable the UART RTS line (set low; read), 0 to disable the
  * UART RTS line (set high; write).
  */
-void dmx_uart_set_rts(uart_dev_t *uart, int set);
+void dmx_uart_set_rts(dmx_port_t dmx_num, int set);
 /**
  * @brief Resets the UART RX FIFO.
  *
  * @param uart A pointer to a UART port.
  */
-void dmx_uart_rxfifo_reset(uart_dev_t *uart);
+void dmx_uart_rxfifo_reset(dmx_port_t dmx_num);
 
 /**
  * @brief Gets the length of the UART TX FIFO.
@@ -164,7 +168,7 @@ void dmx_uart_rxfifo_reset(uart_dev_t *uart);
  * @param uart A pointer to a UART port.
  * @return The length of the UART TX FIFO.
  */
-uint32_t dmx_uart_get_txfifo_len(uart_dev_t *uart);
+uint32_t dmx_uart_get_txfifo_len(dmx_port_t dmx_num);
 
 /**
  * @brief Writes to the UART TX FIFO.
@@ -173,14 +177,14 @@ uint32_t dmx_uart_get_txfifo_len(uart_dev_t *uart);
  * @param buf The source buffer from which to write.
  * @param size The number of bytes to write.
  */
-void dmx_uart_write_txfifo(uart_dev_t *uart, const void *buf, size_t *size);
+void dmx_uart_write_txfifo(dmx_port_t dmx_num, const void *buf, size_t *size);
 
 /**
  * @brief Resets the UART TX FIFO.
  *
  * @param uart A pointer to a UART port.
  */
-void dmx_uart_txfifo_reset(uart_dev_t *uart);
+void dmx_uart_txfifo_reset(dmx_port_t dmx_num);
 
 #ifdef __cplusplus
 }
